+++ frontend/src/components/ProfilUser/ProfilUser.jsx
@@ -1,43 +1,9 @@
 import styles from "../../pages/styles/profile.module.css";
 import { Age } from "../../services/utilities";
-import { useState, useEffect } from "react";
+import DeleteInterestForUser from "../DeleteInterestForUser/DeleteInterestForUser"
 
 const profilUser = ({ user }) => {
-  const [userInterests, setUserInterests] = useState([]);
-  useEffect(()=> {
-    if(user) {
-      setUserInterests(user.interests);
-    }
-   
-  }, [user]);
-
-  const deleteInterest = async (idToDelete) => {
-    try {
-      const response = await fetch(
-        `http://localhost:5000/haveinterests/${user.id}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-            "Authorization": `Bearer${localStorage.getItem("jwtToken")}`,
-          },
-          body: JSON.stringify({interest_id: idToDelete}),
-        }
-      );
- 
-      if (response.status === 204) {
-        const newUserInterest = userInterests.filter(
-          (interet) => interet.id !== idToDelete
-        );
-        setUserInterests(newUserInterest);
-        } else {
-        console.error("Update failed:", response.statusText);
-      }
-    } catch (error) {
-      console.error("Error during update:", error);
-    }
-  };
-
+  
   return (
     <div> {user && (
         <section className="user-profile">
@@ -56,18 +22,7 @@ const profilUser = ({ user }) => {
 
             <h1 className={styles.title1}>Your actual interest</h1>
             <aside>
-              <ul className={styles.listProfil}>
-                {userInterests.length > 0 &&
-                  userInterests.map((interest) => (
-                    <li
-                      className={styles.bubble}
-                      key={interest.id}
-                      onClick={() => deleteInterest(interest.id)}
-                    >
-                      {interest.interest}
-                    </li>
-                  ))}
-              </ul>
+             <DeleteInterestForUser DataUserInterests={user.interests} id={user.id}/>
             </aside>
           </article>
           </section>
